Test.QuickCheck.Instances.List

Documentation

anyList :: Gen a -> Gen [a]

Generates any list (possibly empty) with the contents generated using its argument.

nonEmpty :: Gen a -> Gen [a]

Generates a non-empty list with the contents generated using its argument.

infiniteList :: Gen a -> Gen [a]

Generates an infinite list with contents generated using its argument

setLength :: Int -> Gen a -> Gen [a]

Generates a list with a set length

increasing :: (Arbitrary a, Eq a, Num a) => Gen [a]

Generate increasing towards infinity

nondecreasing :: (Arbitrary a, Num a) => Gen [a]

Generate nondecreasing values

increasingInf :: (Arbitrary a, Eq a, Num a) => Gen [a]

Generate an infinite list of increasing values

nondecreasingInf :: (Arbitrary a, Num a) => Gen [a]

Generate an infinite list of nondecreasing values

decreasing :: (Arbitrary a, Eq a, Num a) => Gen [a]

Generate increasing towards infinity

nonincreasing :: (Arbitrary a, Num a) => Gen [a]

Generate nondecreasing values

decreasingInf :: (Arbitrary a, Eq a, Num a) => Gen [a]

Generate an infinite list of increasing values

nonincreasingInf :: (Arbitrary a, Num a) => Gen [a]

Generate an infinite list of nondecreasing values